Add method BufRegistry::unregister
This supports the io_uring_register facility for buffer deregistration,
allowing different buffer sets to be registered over the lifetime
of an io-uring descriptor.
As long as the old buffer allocations are kept alive by BufRegistry or
FixedBuf instances referring to them, there is no ambiguity for
the kernel when an old buffer instance is erroneously used after
its registration is revoked, even if the buffer index has since been
reused for a newly registered buffer, because the new buffer has to use
a different memory allocation.