Test graceful invalidation of fixed buffers
When BufRegistry::unregister is called, but some FixedBuf handles
obtained from that BufRegistry instance are still around, operations
using these handles are expected to fail gracefully even if the buffer
indices have since been re-registered for new buffers.

This is an edge case, but adding a test for it validates the design
with the capability to unregister the previously registered buffers
and register new ones while in the same io-uring runtime context.