fs: implement `rename`
Implement the `rename` operation for `fs`.
This operation uses `IORING_OP_RENAMEAT` as the opcode for the submitted
SQE.

Both file descriptors are set to the special value `AT_FDCWD` so that
the given paths are interpreted relative to the current working
directory of the calling process.

The operation is submitted with a zero `flags` argument so that it would
be equivalent to the corresponding operation in `tokio:fs`/`std:fs`.

Refs: #48