chore: Tokio-uring design proposal (#1)
[Rendered](https://github.com/tokio-rs/tokio-uring/blob/design-doc/DESIGN.md)

The RFC proposes a new asynchronous Rust runtime backed by io-uring as a new crate: tokio-uring. The API aims to be as close to idiomatic Tokio, but will deviate when necessary to provide full access to io-uring's capabilities. It also will be compatible with existing Tokio libraries. The runtime will use an isolated thread-per-core model, and many types will be `!Send`.

The repo contains a proof-of-concept implementation.