io: refactor post-op logic in ops into Completable
As a first step towards implementing cancellation of in-flight operations, we need to move the post-op logic into the scope of Op so that we can eventually expose that directly instead of hiding it within async blocks. This will allow us to expose a public, async cancel method on it which cancels the in-flight op and awaits its termination.