Fix race condition on multi CQE's
Operations with multi CQE's may post multiple completions
before a waker has been registered with the Lifecycle.

This hit the "entered unreachable code" state in Lifecycle::complete().

To fix this, this commit changes completion to hold completions in a
smallvec, which captures and records all completions that occur before
the first poll occurs. This is safe, as smallvec will grow if needed,
but shouldn't introduce any performance hit for single CQE operations.