rt: fix data race (#146)
Fixes #145. A data race was observed in Pr #144, where with high concurrency in user-space, and single threaded sqpoll in the ring, we could trigger a panic when submitting an entry to the queue

The speculation is that the submission queue is full, but no ops have yet been executed and placed in completion queue by the ring. The submit call therefore submits the queue, but doesn't free any sqe's .

The fix, which is not very elegant, does busy-polling on full.

Co-authored-by: ollie-etl <Oliver Bunting@etlsystems.com>