poll_fn and Unpin: fix pinning
Keeping this crate's poll_fn in sync with that of rust's
library/core/src/future/poll_fn.rs.

Their merge provides all the details.

    https://github.com/rust-lang/rust/pull/102737

None of this crate's uses of poll_fn are affected. It protects against
misuse in the future which could have led to UB as the above mentioned
PR points out with an example of a double free becoming possible.