fs: Remove processing of EINTR
Remove checks for ErrorKind::Interrupted from:
- File::read_exact_at
- File::write_all_at

As mentioned in the implementation of write_all for sockets,
this error code should never be returned with the way the runtime
drives io-uring.
Also update the similar comments in net that needed a bit more
editing since the commented-out match arms were removed.