buf/fixed: re-enable off-runtime initialization
There is no need for the driver when a fixed buffer collection
is populated. The driver comes in when registering and unregistering
the buffers, and then as well, we don't need to preserve the
driver handle.

Change the doc examples to show and exercise off-runtime creation
of FixedBufPool and FixedBufRegistry.