Restore off-runtime creation of fixed buffer collections (#200)
#196 broke the ability to populate a collection of fixed buffers
off-runtime and then have it registered inside a runtime.

There is no need for the driver when a fixed buffer collection is
populated. The driver comes in when registering and unregistering the
buffers, and then as well, we don't currently need the collection to be
affine to the driver. The driver always compares the pointer passed in
to prevent unregistering the wrong buffers, so we can just grab the
driver from the context.

Change the doc examples to show and exercise off-runtime creation of
`FixedBufPool` and `FixedBufRegistry`.