Rewrite FixedBufPool::next to be non-allocating
The wakeup mechanism is now based on tokio::sync::Notify
and wakes up a single task pending for the checked-in capacity
buffer every time.
Remove FixedBufPool::poll_next and the named Next future.
These could not be rewritten with the Notified future without
replicating some of the unsafe magic.
In the future, TAIT can resurrect the Next named type.