Generialise FixedBufRegistry and FixedBufPool
Extends both registries to accept any `Iter` of `IoBufMut`, instead
of Vec<u8>.

This is motivated by my usecase, which cannot safely provide its memory
as a Vec, and needs custom cleanup.

This implementation, which explcitly keeps the backing buffers present,
rather then mem::forget and recreate means the Drop handling is preserved.

As a trivial example, this would now admit Box<[u8]> as a buffer.