fix: API IO bounds should use BoundedBuf
Some of the more recently introduced API IO calls used the original
IoBuf and IoBufMut traits without realizing the more general trait is
BoundedBuf and BoundedBufMut.

Interesting that callers to the library might not notice until they try
to pass a slice of a buffer to one of the IO calls.