add more options for statx
This adds a top level statx function that takes a path obviating the
need to open a file first.

This also adds a builder, named StatxBuilder, to give more control over
the dirfd, path, flags and mask that are ultimately being passed to the
uring statx operation.

The builder can be gotten with or without an open file.

Also add statx tests to the example, test_create_dir_all.