adds pathname option to StatxBuilder
And removes the statx_path(path) terminator to the builder.
The builder api is both simplificied and made more complex.

Simplied because there is just one terminator now, statx().

More complex because the pathname(path) option can fail to create a
CString so needs to return an io::Result<&mut Self> so the builder's
chain of options has this little hiccup in it when setting the path.