fix: rework SharedFd
This allows a 'file.close.wait' to return a true close return value.
And it fixes the problem of indefinitely hanging when the SharedFd was
still held by an in-flight operation.

It simplifies the low level close logic, using the async uring close
operation when the caller is using file.close.wait but using the
synchronous close mechanism provided by the std library when the SharedFd
is allowed to go out of scope without having been closed first.

Manual closing of the file should be encouraged because it allows the
user to see the return code.

In order to let the file owner call close and not hang when an in-flight
was still in progress, this reworks the future to return when it sees
the strong reference count drop to 1.