rt: add new api for oneshot op submission and creation (#244)
This change adds a new API for creating and submitting oneshot
operations. This new API is intended to obsolete the existing system,
however transferring over has been left out of this PR to keep the
change small.

It is intended that a similar API for multishot operations be landed in
a followup PR as well. This was also left out of this PR to keep the
change small.

This refactoring paves the way for further work on SQE linking,
multishot operations, and other improvements and additions.

The goal of this change is to allow us to split up the oneshot and
multishot logic to allow for a cleaner and more extensible system, allow
for user-provided operations, and allow users to control when and how
their ops get submitted to the squeue.